Django_gogs/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
django-redis==5.4.0
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
